requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
pandas==2.1.4
openai==1.3.7
python-dotenv==1.0.0
streamlit==1.28.1
fastapi==0.104.1
uvicorn==0.24.0
thefuzz==0.20.0
python-Levenshtein==0.23.0
googlesearch-python==1.2.3
validators==0.22.0
